# Refactor Solana collection