--- a/blocks/solana/src/lib.rs
+++ b/blocks/solana/src/lib.rs
@@ -1,6 +1,7 @@
 mod account_activity;
 mod blocks;
 mod blocks_without_votes_all;
+mod collect_events;
 mod counters;
 mod events;
 mod instruction_calls;