--- a/blocks/solana/src/transactions.rs
+++ b/blocks/solana/src/transactions.rs
@@ -61,6 +61,54 @@ pub fn collect_transactions(transactions: &Vec<(usize, &ConfirmedTransaction)>, 
     trx_vec
 }
 
+pub fn collect_transaction(transaction: &ConfirmedTransaction, index: usize, block_info: &BlockInfo, timestamp: &BlockTimestamp) -> RawTransaction {
+    let meta = transaction.meta.as_ref().expect("Transaction meta is missing");
+    let trx = transaction.transaction.as_ref().expect("Transaction is missing");
+    let message = trx.message.as_ref().expect("Transaction message is missing");
+    let header = message.header.as_ref().expect("Transaction header is missing");
+
+    let account_keys = get_account_keys_extended(transaction);
+    let success = meta.err.is_none();
+    let err = match &meta.err {
+        Some(err) => decode_transaction_error(&err.err),
+        None => String::new(),
+    };
+
+    let signers = message.account_keys.iter().take(trx.signatures.len()).map(|key| base58::encode(key)).collect::<Vec<String>>();
+    let signers_str = build_csv_string(&signers);
+
+    RawTransaction {
+        block_time: Some(timestamp.time),
+        block_hash: timestamp.hash.clone(),
+        block_date: timestamp.date.clone(),
+        block_slot: block_info.slot,
+        block_height: block_info.height,
+        block_previous_block_hash: block_info.previous_block_hash.clone(),
+        block_parent_slot: block_info.parent_slot,
+        id: transaction.id(),
+        index: index as u32,
+        fee: meta.fee,
+        required_signatures: header.num_required_signatures,
+        required_signed_accounts: header.num_readonly_signed_accounts,
+        required_unsigned_accounts: header.num_readonly_unsigned_accounts,
+        signature: transaction.id(),
+        success,
+        error: err,
+        recent_block_hash: base58::encode(&message.recent_blockhash),
+        account_keys: build_csv_string(&account_keys),
+        log_messages: build_csv_string(&meta.log_messages),
+        // TODO: output as uint array when sink-files supports it
+        pre_balances: build_csv_string(&meta.pre_balances.iter().map(|balance| balance.to_string()).collect::<Vec<String>>()),
+        // TODO: use Array(Text) once sink-files supports it
+        post_balances: build_csv_string(&meta.post_balances.iter().map(|balance| balance.to_string()).collect::<Vec<String>>()),
+        // TODO: use Array(Text) once sink-files supports it
+        signatures: build_csv_string(&trx.signatures.iter().map(|sig| base58::encode(sig)).collect::<Vec<String>>()),
+        signer: message.account_keys.iter().take(trx.signatures.len()).map(|key| base58::encode(key)).next().unwrap(),
+        // TODO: use Array(Text) once sink-files supports it
+        signers: signers_str,
+    }
+}
+
 pub fn decode_transaction_error(err: &Vec<u8>) -> String {
     match TransactionErrorDecoder::decode_error(err) {
         Ok(decoded_error) => TransactionErrorDecoder::format_error(&decoded_error),