Refactored the collection logic of Solana blocks so that we don't iterate through the same `transactions` array multiple times.

Performance difference is hard to tell as the task is connection-bound

For 500 blocks:

## Before refactoring

```bash
make parquet  13.85s user 1.80s system 40% cpu 38.502 total
make parquet  14.34s user 1.81s system 41% cpu 39.327 total
make parquet  14.22s user 1.85s system 41% cpu 39.003 total
```

## After refactoring

```bash
make parquet  14.30s user 1.90s system 40% cpu 39.813 total
make parquet  15.10s user 1.97s system 42% cpu 39.765 total
make parquet  14.84s user 1.94s system 42% cpu 39.550 total
```